Add automatic emails to new email listing tabs
[MAILPOET-2775]
